requirements.txt
asgiref
certifi
charset-normalizer
cloudinary
crispy-bootstrap5
dj-database-url
Django
django-cloudinary-storage
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
idna
Markdown
MouseInfo
numpy
opencv-contrib-python
pillow
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
Pygments
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
requests
six
sqlparse
typing_extensions
tzdata
urllib3